git-svn-id: svn://mmweb.animal.net.cn/nb_icesheet/trunk@58 ba629b7a-d26c-4841-822e-5ac455a76d44

+++ src/Universal/const.h
@@ -9,8 +9,8 @@
 #define UNIVERSAL_CONST_H_
 
 
-const float MATRIX_ZERO = -9998;
-const float NODATA = -9999;
+const int MATRIX_ZERO = -9998;
+const int NODATA = -9999;
 const bool DEBUG = true;
 
 #endif /* UNIVERSAL_CONST_H_ */
